# Faster .rank .select