Hello,

This PR follow-up https://github.com/RoaringBitmap/RoaringBitmap/issues/191

The general idea is to cache cardinalities, cumulated through buckets. This yield better performance when doing a lot of .rank and .select operations over a RoaringBitmap which is not written anymore.

Raw performance results (see TestRoaringBitmap_FastRank):

    AddOnlyThenSelectOnly FastRankRoaringBitmap=207ms
    AddOnlyThenSelectOnly RoaringBitmap=6064ms

    AddSelectRemoveAreMixed FastRankRoaringBitmap=1055ms
    AddSelectRemoveAreMixed RoaringBitmap=375ms


I am not fully-convinced `FastRankRoaringBitmap` is a proper naming for this class