@blasd 2017-12-07
I changed the way to mark the cardianlity cache as dismissed. It enables re-using the int[] used as cache. The performance seems not improved but it will lower the GC pressure when doing a lot of consecutive .add/.select

    AddSelectRemoveAreMixed LargeInts FastRankRoaringBitmap=937ms
    AddSelectRemoveAreMixed LargeInts RoaringBitmap=387ms
    AddSelectRemoveAreMixed SmallInts FastRankRoaringBitmap=732ms
    AddSelectRemoveAreMixed SmallInts RoaringBitmap=625ms
    AddOnlyThenSelectOnly FastRankRoaringBitmap=184ms
    AddOnlyThenSelectOnly RoaringBitmap=5696ms