@lemire 2017-12-07
@blasd 

I understand what you did and why you did it, and I am fine with merging it if you recommend merger. However, I have to ask...

I find it awkward to use a special value located in the first component of the vector to determine whether the cache is valid. It is not entirely intuitive, though it is clear enough.

Couldn't we, more simply, have a boolean flag. 

So instead of 

```
highToCumulatedCardinality == null || highToCumulatedCardinality[0] == DISMISSED_MARKER;
```

just have...

```
boolean isCacheDismissed;
```

The downside is that we add about a word of storage to store the boolean value, but that's probably negligible. It may even end up improving performance in that it is cheaper to check one flag than doing an array access.

I don't think it is a big deal, but it is a matter of elegance/simplicity.