@blasd 2017-12-13
Hello,
I am completely fine to rely on a dedicated boolean.

Current implementation extends RoaringBitmap. Then, it does not work with `ImmutableRoaringBitmap`. It would require much less duplicated code to handle both case (and possibly part of `Roaring64NavigableMap`) if `RoaringArray` and `PointableRoaringArray` shared a common interface (e.g. `getContainerAtIndex(int i)` ; and the same for `Container.getCardinality and MappeableContainer.getCardinality()`). These additional interfaces would enable `FastRankRoaringBitmap `to be a decorator over underlying XXX (XXX being another interfaces as I guess would we not add `.getContainerAtIndex` to `ImmutableBitmapDataProvider`).

As these modifications requires many changes to existing interfaces (or changes to existing one), I guess it should be done in a later step.