@lemire 2017-12-13
Ok. So let us be clear: you recommend as-is merge and release... but you’ll consider improvements later.

Is that correct?