@blasd 2017-12-13
It can be merged as-is. I am not fully satisfied with the naming but if you
are ok with it, we are good to go.

I am also ready to try to implement the suggested improvments (soon enough
to differ merging) but I need your feeling about it. Anyway, even if you
are ok with the suggested refactoring, it may be preferable to do first an
intermediate (but still functional) release.

Le 13 déc. 2017 2:05 PM, "Daniel Lemire" <notifications@github.com> a
écrit :

Ok. So let us be clear: you recommend as-is merge and release... but you’ll
consider improvements later.

Is that correct?

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub
<https://github.com/RoaringBitmap/RoaringBitmap/pull/192#issuecomment-351385282>,
or mute the thread
<https://github.com/notifications/unsubscribe-auth/ACBRF93Wt0z5nU_MOR8394EvnHqRL9uVks5s_8twgaJpZM4Q5MBm>
.