@blasd 2017-12-13
The introduction of described interfaces seems somewhat related to https://github.com/RoaringBitmap/RoaringBitmap/issues/196, as they may help mixing different Containers implementation inside some `HybridRoaringBitmap `