@lemire 2017-12-13
I'll merge and release soon.