@lemire 2017-12-19
Merged. Sorry for the delay.